Refactor model config script
